support dynamic function calls in component model
This addresses #4310, introducing a new `component::values::Val` type for
representing component values dynamically.  It also adds a `call` method to
`component::func::Func`, which takes a slice of `Val`s as parameters and returns
a `Result<Val>` representing the result.

As an implementation detail, I've also added a `component::values::Type` type,
which is an owned, despecialized version of
`wasmtime_environ::component::InterfaceType`.  That serves two purposes:

- It allows us to despecialize as the first step, which reduces the number of cases to consider when typechecking and lowering.
- It avoids needing to borrow the store both mutably and immutably when lowering, as we would if we used `InterfaceType`s.

Finally, I happened to notice that the `ComponentType::SIZE32` calculation in
`expand_record_for_component_type` needed a correction, so I did that.

Signed-off-by: Joel Dice <joel.dice@fermyon.com>