add `StreamReader::try_into` and use it in `Response::into_http_with_getter`
This allows the embedder to consume a host-created `StreamReader` and extract
the `StreamProducer` (or an `Unpin` subset of it) used to create it.  This
enables "short-circuiting" host-to-host streams, bypassing the guest entirely.

Concretely, it is now possible to dispose of a `wasi:http/handler@0.3.x` guest
and its store before the request and/or response bodies have finished streaming
in the case where both the read and write ends of the body stream are owned by
the host.  In this case, the inbound body is reused as the outbound body with no
additional processing or memory overhead.

Signed-off-by: Joel Dice <joel.dice@fermyon.com>