reset read/write state back to `Open` on event delivery
If one end of a stream does a partial read or write, we leave the other end in a
`GuestReady` state, allowing further reads or writes to proceed until the buffer
has been drained or filled, respectively.  However, once we've delivered the
event regarding the partial operation, we need to set the state back to `Open`,
since we'll have released the buffer back to the guest at that point.

Signed-off-by: Joel Dice <joel.dice@fermyon.com>